requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
flask==3.0.2
waitress==3.0.2
apscheduler==3.10.4
feedparser==6.0.11
readability-lxml==0.8.4.1
Pillow==11.3.0
bleach>=6.0.0
curl_cffi>=0.7.0
